Rename test/tests/ → test/cases/, AsciiSnapshot → ascii-output-fixture
- test/tests/ stutters; test/cases/ reads more naturally as the layer
  containing runnable test suites
- AsciiSnapshot.js renamed to ascii-output-fixture.js: more literal,
  avoids snapshot-test update semantics implied by "snapshot"

Addresses review feedback from @JamesCropcho on PR #285.
