+++ test/cases/analysis/BasicAnalysisTest.js
@@ -5,7 +5,7 @@ import VarietyHarness from '../../helpers/VarietyHarness.js';
 const test = new VarietyHarness('test', 'users');
 
 import sampleData from '../../fixtures/seed-data.js';
-import expectedAscii from '../../fixtures/AsciiSnapshot.js';
+import expectedAscii from '../../fixtures/ascii-output-fixture.js';
 
 describe('Basic Analysis', () => {
