+++ test/cases/analysis/DatatypeRecognitionTest.js
