+++ test/cases/analysis/EmptyResultsTest.js
