+++ test/cases/analysis/LimitResultsAnalysisTest.js
