+++ test/cases/analysis/MaxDepthAnalysisTest.js
