+++ test/cases/analysis/PrototypeNamedKeysTest.js
