+++ test/cases/analysis/QueryLimitedAnalysisTest.js
