+++ test/cases/analysis/ShowArrayElementsTest.js
