+++ test/cases/analysis/SortedAnalysisTest.js
