+++ test/cases/analysis/UnnamedObjectsAnalysisTest.js
