+++ test/cases/cli/BinWrapperTest.js
