+++ test/cases/cli/CliOptionsTest.js
