+++ test/cases/mongo-shell/AdapterValidationTest.js
