+++ test/cases/persistence/ContinuousLoggingTest.js
