+++ test/cases/plugins/PluginTest.js
