+++ test/fixtures/ascii-output-fixture.js
